Fantasy General — Syntax update
* Update script to 2.11 syntax
* Update authors list
* Fix language-based package name
* Fix dependencies list
* Fix script execution in dry run mode

+++ play.it-2/games/play-fantasy-general.sh
@@ -1,8 +1,9 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
 # Copyright (c) 2015-2019, Antoine "vv221/vv222" Le Gonidec
+# Copyright (c) 2016-2019, Solène "Mopi" Huault
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -30,11 +31,11 @@ set -o errexit
 
 ###
 # Fantasy General
-# build native Linux packages from the original installers
+# build native packages from the original installers
 # send your bug reports to vv221@dotslashplay.it
 ###
 
-script_version=20180224.1
+script_version=20190721.1
 
 # Set game-specific variables
 
@@ -55,14 +56,14 @@ ARCHIVE_GOG_FR_MD5='1b188304b4cca838e6918ca6e2d9fe2b'
 ARCHIVE_GOG_FR_VERSION='1.0-gog2.0.0.8'
 ARCHIVE_GOG_FR_SIZE='240000'
 
-ARCHIVE_DOC1_PATH='data/noarch/docs'
-ARCHIVE_DOC1_FILES='./*'
+ARCHIVE_DOC0_MAIN_PATH='data/noarch/docs'
+ARCHIVE_DOC0_MAIN_FILES='*'
 
-ARCHIVE_DOC2_PATH='data/noarch/data'
-ARCHIVE_DOC2_FILES='./*.txt'
+ARCHIVE_DOC1_MAIN_PATH='data/noarch/data'
+ARCHIVE_DOC1_MAIN_FILES='*.txt'
 
-ARCHIVE_GAME_PATH='data/noarch/data'
-ARCHIVE_GAME_FILES='./*'
+ARCHIVE_GAME_MAIN_PATH='data/noarch/data'
+ARCHIVE_GAME_MAIN_FILES='*'
 
 GAME_IMAGE='./game.ins'
 
@@ -72,55 +73,66 @@ CONFIG_FILES='dat/prefs.dat'
 APP_MAIN_TYPE='dosbox'
 APP_MAIN_EXE='exe/barena.exe'
 APP_MAIN_ICON='data/noarch/support/icon.png'
-APP_MAIN_ICON_RES='256'
 
 PACKAGES_LIST='PKG_MAIN'
 
 PKG_MAIN_ID="$GAME_ID"
-PKG_MAIN_ID_EN="${GAME_ID}-en"
-PKG_MAIN_ID_FR="${GAME_ID}-fr"
+PKG_MAIN_ID_GOG_EN="${GAME_ID}-en"
+PKG_MAIN_ID_GOG_FR="${GAME_ID}-fr"
 PKG_MAIN_PROVIDE="$PKG_MAIN_ID"
-PKG_MAIN_ARCH='32'
-PKG_MAIN_DEPS_DEB="$PKG_DATA_ID, dosbox"
-PKG_MAIN_DEPS_ARCH="$PKG_DATA_ID dosbox"
+PKG_MAIN_DEPS='dosbox'
 
 # Load common functions
 
-target_version='2.4'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
-#shellcheck source=play.it-2/lib/libplayit2.sh
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
 tolower "$PLAYIT_WORKDIR/gamedata"
+prepare_package_layout
 
-organize_data 'DOC1'      "$PATH_DOC"
-organize_data 'DOC2'      "$PATH_DOC"
-organize_data 'GAME' "$PATH_GAME"
-get_icon_from_temp_dir 'APP_MAIN'
+# Get icon
 
+icons_get_from_workdir 'APP_MAIN'
 rm --recursive "$PLAYIT_WORKDIR/gamedata"
 
 # Write launchers
 
-write_launcher 'APP_MAIN'
+launchers_write 'APP_MAIN'
+
+# Run game binary from its direct parent directory
 
 # shellcheck disable=SC2016
-sed --in-place 's|$APP_EXE \($APP_OPTIONS $@\)|cd ${APP_EXE%/*}\n${APP_EXE##*/} \1|' "${PKG_MAIN_PATH}${PATH_BIN}/$GAME_ID"
+pattern='s|$APP_EXE \($APP_OPTIONS $@\)|cd ${APP_EXE%/*}\n${APP_EXE##*/} \1|'
+file="${PKG_MAIN_PATH}${PATH_BIN}/$GAME_ID"
+if [ $DRY_RUN -eq 0 ]; then
+	sed --in-place "$pattern" "$file"
+fi
 
 # Build package
